Refactor cicd test pipeline for more immediate results.
