+++ src/setup/integration-test/test.yml
@@ -1,0 +1,20 @@
+service: TestService
+frameworkVersion: "3"
+provider:
+    name: aws
+    runtime: python3.9
+    region: us-west-1
+package:
+    individually: true
+functions:
+    testFunction1:
+        handler: hellopy/lambda_function.lambda_handler
+        runtime: python3.9
+        name: parallelism1_0_0
+        events:
+            - httpApi:
+                path: /parallelism1_0_0
+                method: GET
+        package:
+            patterns:
+                - hellopy/lambda_function.py
